Create module for SD card.
