+++ libraries/common-arm/src/sd/mod.rs
@@ -1,0 +1,2 @@
+pub mod sd_interface;
+pub use sd_interface::SdInterface;
